Move setting widget prop types into core
ChartSettingEnumToggleProps and ChartSettingSegmentedControlProps (the two
type-only component imports in lib/settings/graph.ts) now live in
types/widget-props.ts; the widget components import them from there, so the
type edge runs UI to core. Type-only imports are real edges under
tsPreCompilationDeps, so this affects the affected-tests planner too.
