+++ frontend/src/metabase/visualizations/components/settings/ChartSettingEnumToggle.tsx
@@ -1,12 +1,5 @@
 import { Switch } from "metabase/ui";
-
-export type ChartSettingEnumToggleProps<T extends string> = {
-  value: T | undefined;
-  onChange: (value: T) => void;
-  id?: string;
-  checkedValue: T;
-  uncheckedValue: T;
-};
+import type { ChartSettingEnumToggleProps } from "metabase/visualizations/types/widget-props";
 
 export const ChartSettingEnumToggle = <T extends string>({
   value,
